+++ test/point.test.js
@@ -1,4 +1,4 @@
-var assert = require('assert')
+var assert= require('assert')
 
 var BigInteger = require('bigi')
 
@@ -6,8 +6,6 @@ var ecurve = require('../')
 var ECCurveFp = ecurve.ECCurveFp
 var ECPointFp = ecurve.ECPointFp
 var getECParams = ecurve.getECParams
-
-require('terst')
 
 var fixtures = require('./fixtures/point')
 
@@ -77,9 +75,9 @@ describe('ECPointFp', function() {
       })
     })
 
-    describe('> when compressed flag is set to true', function() {
+    describe('> when compressed flag is seassertto true', function() {
       describe('> when false is passed', function() {
-        it('should return encoded (not compressed)', function() {
+        it('should return encoded (noassertcompressed)', function() {
           var x = "55066263022277343669578718895168534326250603453777594175500187360389116729240"
           var y = "32670510020758816978083085130507043184471273380659243275938904335757337482424"
           var res = "0479be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f81798483ada7726a3c4655da4fbfc0e1108a8fd17b448a68554199c47d08ffb10d4b8"
@@ -89,12 +87,12 @@ describe('ECPointFp', function() {
           var Q = new ECPointFp(curve, curve.fromBigInteger(new BigInteger(x)), curve.fromBigInteger(new BigInteger(y)))
           Q.compressed = true
           var encoded = Q.getEncoded(doCompress)
-          EQ (encoded.toString('hex'), res)
+          assert.equal(encoded.toString('hex'), res)
         })
       })
 
       describe('> when true is passed', function() {
-        it('should return encoded (not compressed)', function() {
+        it('should return encoded (noassertcompressed)', function() {
           var x = "55066263022277343669578718895168534326250603453777594175500187360389116729240"
           var y = "32670510020758816978083085130507043184471273380659243275938904335757337482424"
           var res = "0279be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f81798"
@@ -104,14 +102,14 @@ describe('ECPointFp', function() {
           var Q = new ECPointFp(curve, curve.fromBigInteger(new BigInteger(x)), curve.fromBigInteger(new BigInteger(y)))
           Q.compressed = true
           var encoded = Q.getEncoded(doCompress)
-          EQ (encoded.toString('hex'), res)
+          assert.equal(encoded.toString('hex'), res)
         })
       })
     })
 
-    describe('> when compressed flag is set to false', function() {
+    describe('> when compressed flag is seassertto false', function() {
       describe('> when false is passed', function() {
-        it('should return encoded (not compressed)', function() {
+        it('should return encoded (noassertcompressed)', function() {
           var x = "55066263022277343669578718895168534326250603453777594175500187360389116729240"
           var y = "32670510020758816978083085130507043184471273380659243275938904335757337482424"
           var res = "0479be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f81798483ada7726a3c4655da4fbfc0e1108a8fd17b448a68554199c47d08ffb10d4b8"
@@ -121,12 +119,12 @@ describe('ECPointFp', function() {
           var Q = new ECPointFp(curve, curve.fromBigInteger(new BigInteger(x)), curve.fromBigInteger(new BigInteger(y)))
           Q.compressed = false
           var encoded = Q.getEncoded(doCompress)
-          EQ (encoded.toString('hex'), res)
+          assert.equal(encoded.toString('hex'), res)
         })
       })
 
       describe('> when true is passed', function() {
-        it('should return encoded (not compressed)', function() {
+        it('should return encoded (noassertcompressed)', function() {
           var x = "55066263022277343669578718895168534326250603453777594175500187360389116729240"
           var y = "32670510020758816978083085130507043184471273380659243275938904335757337482424"
           var res = "0279be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f81798"
@@ -136,14 +134,14 @@ describe('ECPointFp', function() {
           var Q = new ECPointFp(curve, curve.fromBigInteger(new BigInteger(x)), curve.fromBigInteger(new BigInteger(y)))
           Q.compressed = false
           var encoded = Q.getEncoded(doCompress)
-          EQ (encoded.toString('hex'), res)
+          assert.equal(encoded.toString('hex'), res)
         })
       })
     })
 
     describe('> when getEncoded() has no parameter', function() {
-      describe('> when compressed flag is set to false', function() {
-        it('should return encoded (not compressed)', function() {
+      describe('> when compressed flag is seassertto false', function() {
+        it('should return encoded (noassertcompressed)', function() {
           var x = "55066263022277343669578718895168534326250603453777594175500187360389116729240"
           var y = "32670510020758816978083085130507043184471273380659243275938904335757337482424"
           var res = "0479be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f81798483ada7726a3c4655da4fbfc0e1108a8fd17b448a68554199c47d08ffb10d4b8"
@@ -152,12 +150,12 @@ describe('ECPointFp', function() {
           var Q = new ECPointFp(curve, curve.fromBigInteger(new BigInteger(x)), curve.fromBigInteger(new BigInteger(y)))
           Q.compressed = false
           var encoded = Q.getEncoded()
-          EQ (encoded.toString('hex'), res)
+          assert.equal(encoded.toString('hex'), res)
         })
       })
 
-      describe('> when compressed flag is set to true', function() {
-        it('should return encoded (not compressed)', function() {
+      describe('> when compressed flag is seassertto true', function() {
+        it('should return encoded (noassertcompressed)', function() {
           var x = "55066263022277343669578718895168534326250603453777594175500187360389116729240"
           var y = "32670510020758816978083085130507043184471273380659243275938904335757337482424"
           var res = "0279be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f81798"
@@ -166,7 +164,7 @@ describe('ECPointFp', function() {
           var Q = new ECPointFp(curve, curve.fromBigInteger(new BigInteger(x)), curve.fromBigInteger(new BigInteger(y)))
           Q.compressed = true
           var encoded = Q.getEncoded()
-          EQ (encoded.toString('hex'), res)
+          assert.equal(encoded.toString('hex'), res)
         })
       })
     })
@@ -188,11 +186,11 @@ describe('ECPointFp', function() {
       var y2 = BigInteger.fromHex("483ADA7726A3C4655DA4FBFC0E1108A8FD17B448A68554199C47D08FFB10D4B8")
       var G2 = new ECPointFp(curve, curve.fromBigInteger(x2), curve.fromBigInteger(y2))
 
-      T (G1.equals(G2))
-      T (G2.equals(G1))
+      assert(G1.equals(G2))
+      assert(G2.equals(G1))
     })
 
-    it('should return false when points are not equal', function() {
+    it('should return false when points are noassertequal', function() {
       var x1 = BigInteger.fromHex("79BE667EF9DCBBAC55A06295CE870B07029BFCDB2DCE28D959F2815B16F81798")
       var y1 = BigInteger.fromHex("483ADA7726A3C4655DA4FBFC0E1108A8FD17B448A68554199C47D08FFB10D4B8")
       var G1 = new ECPointFp(curve, curve.fromBigInteger(x1), curve.fromBigInteger(y1))
@@ -201,8 +199,8 @@ describe('ECPointFp', function() {
       var y2 = BigInteger.fromHex("483ADA7726A3C4655DA4FBFC0E1108A8FD17B448A68554199C47D08FFB10D4B8")
       var G2 = new ECPointFp(curve, curve.fromBigInteger(x2), curve.fromBigInteger(y2))
 
-      F (G1.equals(G2))
-      F (G2.equals(G1))
+      assert(!G1.equals(G2))
+      assert(!G2.equals(G1))
     })
   })
 })